Move client, settings, and setup modules to separate 'storage' subpackage
