+++ pyproject.toml
@@ -116,7 +116,7 @@ output-format = 'grouped'
 target-version = 'py311'
 
 [tool.ruff.lint]
-select = ['B', 'C4', 'C90', 'E', 'F']
+select = ['B', 'C4', 'C90', 'E', 'F', 'I']
 ignore = ['B023']
 
 [tool.ruff.format]
